nc - remove all front end changes implemented by Rowan and Rag (go in separate issue)
