+++ javascript/src/main/pages/Admin/ViewLogins.js
@@ -28,8 +28,7 @@ const ViewLogins = () => {
   return (
     <>
       <h2 className="text-left">Login History</h2>
-      <Button className={"mb-2"}><CSVLink style={{color: "white"}} headers={headers} data={loginTable || []} filename = {"LoginTable.csv"}>Download CSV</CSVLink></Button>
-      {loginTable && <p className="text-left">Number of logins: {loginTable.length}</p>}
+      <Button><CSVLink style={{color: "white"}} headers={headers} data={loginTable || []} filename = {"LoginTable.csv"}>Download CSV</CSVLink></Button>
       <LoginTable loginTable={loginTable || []} admin={true}/>
     </>
   );
