+++ javascript/src/main/pages/TutorAssignment/TutorAssignments.js
@@ -7,7 +7,7 @@ import Loading from "main/components/Loading/Loading";
 import TutorAssignmentTable from "main/components/TutorAssignment/TutorAssignmentTable"
 
 import {useHistory} from "react-router-dom";
-import { CSVLink } from "react-csv";
+
 
 const TutorAssignment = () => {
   const history = useHistory();
@@ -33,29 +33,6 @@ const TutorAssignment = () => {
 
   const newTutorAssignmentButton = <Button className="mb-3" onClick={()=>history.push("/tutorAssignments/new")}>New Tutor Assignment</Button>;
 
-  const headers = [{
-    key: 'id',
-    label: 'Course Id'
-  }, {
-    key: 'course.name',
-    label: 'Course Name'
-  }, {
-    key: 'course.quarter',
-    label: 'Course Quarter'
-  }, {
-    key: 'tutor.firstName',
-    label: 'Tutor First Name'
-  }, {
-    key: 'tutor.lastName',
-    label: 'Tutor Last Name'
-  }, {
-    key: 'tutor.email',
-    label: 'Tutor Email'
-  }, {
-    key: 'assignmentType',
-    label: 'Assignment Type'
-  }];
-
   if (error) {
     return (
       <>
@@ -70,10 +47,7 @@ const TutorAssignment = () => {
 
   return (
     <>
-      {isInstructor && newTutorAssignmentButton}   
-      <br></br>   
-      <Button><CSVLink style={{color: "white"}} headers={headers} data={tutorAssignmentList} filename = {"TutorAssignments.csv"}>Download CSV</CSVLink></Button>
-      <hr></hr>
+      {isInstructor && newTutorAssignmentButton}
       <TutorAssignmentTable tutorAssignments={tutorAssignmentList} isInstructor={isInstructor}/>
     </>
   );
