+++ javascript/src/main/services/selectorSupport.js
@@ -3,6 +3,5 @@ import { fetchWithoutToken } from "main/utils/fetch";
 const quarterProvider = () => fetchWithoutToken("/api/public/quarters");
 const courseProvider = quarter => fetchWithoutToken(`/api/public/courses/forQuarter/${quarter}`);
 const tutorAssignmentProvider = course => fetchWithoutToken(`/api/public/tutorAssignment/${course}`);
-const roomSlotProvider = () => fetchWithoutToken("/api/public/roomslot");
 
-export {quarterProvider, courseProvider, tutorAssignmentProvider, roomSlotProvider };
+export {quarterProvider, courseProvider, tutorAssignmentProvider };
