--- javascript/src/main/utils/RoomSlotTableHelpers.js
@@ -1,15 +1,0 @@
-// Adapted from StackOverflow:
-// https://stackoverflow.com/questions/13898423/javascript-convert-24-hour-time-of-day-string-to-12-hour-time-with-am-pm-and-no
-const toAMPMFormat = (time) => {
-  // Check correct time format and split into components
-  time = time.substring(0, 5).toString ().match (/^([01]\d|2[0-3])(:)([0-5]\d)(:[0-5]\d)?$/) || [time];
-
-  if (time.length > 1) { // If time format correct
-    time = time.slice (1);  // Remove full string match value
-    time[5] = +time[0] < 12 ? 'AM' : 'PM'; // Set AM/PM
-    time[0] = +time[0] % 12 || 12; // Adjust hours
-  }
-  return time.join (''); // return adjusted time or original string
-}
-
-export { toAMPMFormat };
