+++ javascript/src/test/services/OfficeHours/OfficeHourService.test.js
@@ -1,4 +1,4 @@
-import { buildCreateOfficeHour, buildDeleteOfficeHour, buildUpdateOfficeHour } from "main/services/OfficeHours/OfficeHourService";
+import { buildCreateOfficeHour, buildDeleteOfficeHour } from "main/services/OfficeHours/OfficeHourService";
 
 import { fetchWithToken } from "main/utils/fetch";
 
@@ -34,18 +34,7 @@ describe("OfficeHourService tests", () => {
         expect(onError).toBeCalledTimes(1);
     });
 
-    test("buildUpdateOfficeHour and invoke updateOfficeHour", async () => {
-        const updateOfficeHour = buildUpdateOfficeHour(getToken, onSuccess, onError);
-        await updateOfficeHour();
-        expect(onSuccess).toBeCalledTimes(1);
-    });
 
-    test("buildUpdateOfficeHour where we expect onError to be called", async () => {
-        fetchWithToken.mockImplementation( async () => { throw new Error("mock error"); } );
-        const updateOfficeHour = buildUpdateOfficeHour(getToken, onSuccess, onError);
-        await updateOfficeHour();
-        expect(onError).toBeCalledTimes(1);
-    });
 
     test("buildDeleteOfficeHour where we expect onError to be called", async () => {
         fetchWithToken.mockImplementation( async () => { throw new Error("mock error"); } );
