+++ javascript/src/test/utils/OfficeHourFormHelpers.test.js
@@ -1,7 +1,22 @@
-import { checkZoomRoomLink } from "main/utils/OfficeHourFormHelpers"; 
+import { checkTime, checkZoomRoomLink, checkFilled } from "main/utils/OfficeHourFormHelpers"; 
 
 describe("Office Hours form helpers test", () => {
-    
+    describe("checkTime", () => {
+        test("good time", () => {
+            expect(checkTime("1:00PM")).toBe(true);
+            expect(checkTime("12:00PM")).toBe(true);
+            expect(checkTime("9:30AM")).toBe(true);
+            expect(checkTime("11:35AM")).toBe(true);
+        });
+
+        test("bad time", () => {
+            expect(checkTime("01:00PM")).toBe(false);
+            expect(checkTime("1:75PM")).toBe(false);
+            expect(checkTime("S2020")).toBe(false);
+            expect(checkTime("19PM")).toBe(false);
+        });
+    });
+
     describe("checkZoomRoomLink", () => {
         test("good zoom room link", () => {
             expect(checkZoomRoomLink("https://ucsb.zoom.us")).toBe(true);
@@ -19,5 +34,15 @@ describe("Office Hours form helpers test", () => {
         });
     }); 
 
-     
+    describe("checkFilled", () => {
+        test("filled", () => {
+            expect(checkFilled("https://ucsb.zoom.us")).toBe(true);
+            expect(checkFilled("W")).toBe(true);
+            expect(checkFilled("10:30AM")).toBe(true);
+        });
+
+        test("not filled", () => {
+            expect(checkFilled()).toBe(false);
+        });
+    }); 
 })
